Organize Fastmail-specific models, methods under fastmail/
